+++ CesiumGltf/include/CesiumGltf/Model.h
@@ -1,7 +1,8 @@
 #pragma once
 
-#include "CesiumGltf/Library.h"
-#include "CesiumGltf/ModelSpec.h"
+#include <CesiumGltf/Library.h>
+#include <CesiumGltf/ModelSpec.h>
+#include <CesiumUtility/ErrorList.h>
 
 #include <glm/mat4x4.hpp>
 
@@ -22,7 +23,47 @@ struct CESIUMGLTF_API Model : public ModelSpec {
    *
    * @param rhs The model to merge into this one.
    */
-  void merge(Model&& rhs);
+  CesiumUtility::ErrorList merge(Model&& rhs);
+
+  /**
+   * @brief A callback function for {@link forEachPrimitiveInScene}.
+   */
+  typedef void ForEachRootNodeInSceneCallback(Model& gltf, Node& node);
+
+  /**
+   * @brief A callback function for {@link forEachPrimitiveInScene}.
+   */
+  typedef void
+  ForEachRootNodeInSceneConstCallback(const Model& gltf, const Node& node);
+
+  /**
+   * @brief Apply the given callback to the root nodes of the scene.
+   *
+   * If the given `sceneID` is non-negative and exists in the given glTF,
+   * then the given callback will be applied to all nodes of this scene.
+   *
+   * If the given `sceneId` is negative, then the nodes that the callback
+   * will be applied to depends on the structure of the glTF model:
+   *
+   * * If the glTF model has a default scene, then it will
+   *   be applied to all nodes of the default scene.
+   * * Otherwise, it will be applied to all nodes of the the first scene.
+   * * Otherwise (if the glTF model does not contain any scenes), it will
+   *   be applied to the first node.
+   * * Otherwise (if there are no scenes and no nodes), then this method will do
+   *   nothing.
+   *
+   * @param sceneID The scene ID (index)
+   * @param callback The callback to apply
+   */
+  void forEachRootNodeInScene(
+      int sceneID,
+      std::function<ForEachRootNodeInSceneCallback>&& callback);
+
+  /** @copydoc Gltf::forEachRootNodeInScene() */
+  void forEachRootNodeInScene(
+      int sceneID,
+      std::function<ForEachRootNodeInSceneConstCallback>&& callback) const;
 
   /**
    * @brief A callback function for {@link forEachPrimitiveInScene}.